Version 1.8.0
Version 1.8.0 - August 9, 2016
- Added: PreCacheDocument method. Normally the source document is
converted to a special web-friendly format and cached
after the document viewer is displayed in the page and when the document
is viewed for the first time. With this method,
you can do the conversion and caching beforehand so this way, when a
user opens the document for the first time, there
will be no waiting and the document will be loaded immediately from the
cache.
- Improved: Updated document conversion engine.