Version 2.1.0
## Version 2.1.0 - October 26, 2016

- **Added:** Full CAD support. DWG files (AutoCAD R13, R14, 2000, 2004,
2007, 2010, 2013, 2016) and
DXF files (AutoCAD R12, R13, R14, 2000, 2004, 2007, 2010, 2013, 2016)
can now be opened with
DocumentViewer and can now be converted perfectly to Pdf, Ps, Svg, Xaml,
Bmp, Gif, Tif, Png, Jpg
with DocumentConverter. In previous versions, there was DWG and DXF
support but it was pretty
limited and it could only load very few AutoCAD format versions. Now any
AutoCAD file you throw
at it will work.

- **Improved:** DocumentViewer could sometimes show "XPZ invalid part"
error, this happened because
of browser cache conflicting with a new version of the cached file. For
example if you deleted
the cache file and forced regeneration or if a conversion error occured
before, this temporary
but misleading error was shown (it would disappear when you opened the
document second time).
From now on you should never receive this error because we splitted
document conversion and
downloading so incorrect browser caching will be prevented.

- **Improved:** DocumentViewer will show an additional downloading
indicator animation when opening
the document. As you know, upon first time view "please wait a moment"
dialog is displayed.
However even the conversion is completed or when you view the document
second time from cache,
if you have a very slow network and a large document, there can be a
noticable delay until you
see the first page appear. This can feel like the DocumentViewer is not
responsive however it's
actually downloading partially the first page of the document. The
downloading indicator in the
middle will ensure the user that the DocumentViewer is actually busy and
that it will successfully
display the first page in a second or so.

- **Fixed:** When using non-physical DocumentSource (bytes or stream)
and if the source file was a PDF
file, the download button would fail with "Path cannot be null" error.
This occured because the source
copy in cache folder was being overwritten by the generated PDF file so
the download button could not
find the original source copy anymore (even keepSourceCached was set to
true). From now on source copy
will use a different key so that it never overlaps with the conversion
target that have the same
extension (PDF in this case).

- **Improved:** More descriptive errors will be shown for some cases in
DocumentViewer, for example
"Download" and "Download As PDF" buttons and PDF Conversion step will
tell exactly if the error
was due to missing source copy in cache folder or non-existing physical
source file.

- **Fixed:** Trial watermark was being applied twice for PDF source
files (could be understood from the
boldness of red text color). This occured because the watermark was
first applied to the source copy
in cache folder and then applied to the generated intermediate PDF file.
From now on, source copy will
never be touched after being copied from bytes or stream.