Version 2.3.2
## Version 2.3.2 - February 16, 2017

- **Fixed:** Specific languages like pt-BR (not neutral languages like
pt) were not being loaded. New translations under
App_GlobalResources\WebViewer\Html5\Resources\i18n were not picked up.
Changed the translation file name format from
translation-xx_XX.json to translation-xx-XX.json (replaced underscore
with dash) so that it exactly matches .NET culture
names. DocumentViewer.Language property will now throw an error if set
to an invalid .NET culture name.

- **Fixed:** DocumentFormatGroupInfo static constructor error (object
reference not set...) when used before accessing
DocumentFormatInfo.

- **Fixed:** Incorrect "running in demo mode" message when used in
Console Applications even when the license was set.

- **Improved**: Optimized client resources for DocumentViewer.

## Version 2.3.0 - February 13, 2017

- **Fixed:** Bookmarks (Outlines) for PDFs were lost/missing in
DocumentViewer. This also happened with DocumentConverter
for Pdf-to-Pdf conversions.

- **Fixed:** Setting LicenseKey via code, did not remove trial
restrictions if application was not restarted,
e.g. if the code is changed in .aspx or or .asax markup this would not
cause the whole application recompile & restart.
Setting LicenseKey via appSettings in web.config worked because this
already forces application restart.

- **Fixed:** Setting DocumentUltimateWebConfiguration.CacheMaxDays to a
value 0 or less will not cause error during
caching a file (e.g. source copy not found). Current cache subfolder
will be excluded from cache trim operation
so during caching a file, related files for cache key will not be
deleted mistakenly.

- **Fixed:** Possible "Invalid XPZ file: Zip end header data is wrong
size" error on consecutive views if the first view
resulted in a failed conversion. This occured because DocumentCache was
not deleting ghost zero-size files as expected
after a conversion error. This confused the user to think there is a JS
error but the actual error is conversion error
upon first view of the document.

- **Fixed:** Watermarks were not applied to AutoCad files (dxf, dwg) in
DocumentViewer or when converting from
AutoCad to PDF in DocumentConverter.

- **Added**: FastWebViewEnabled property to PdfOutputOptions. By default
DocumentConverter (and so DocumentViewer)
generated Fast Web View enabled PDFs (Linearized PDFs), now this is
optional.

- **Improved**: Updated DocumentViewer client rendering engine. This may
also fix a Chrome v55 rendering glitch:
For some documents, you could see some jumbled text on the upper left
corner of first page. However the glitch was
not permanent, it occured on the same document occasionally. The problem
was not replicated in Chrome v56 though
so it could just be quirk of Chrome v55 only. With updated
DocumentViewer client rendering engine, you should not
see the problem even in Chrome v55 (or earlier) now. Also in
DocumentViewer, changed color of text selection handles.

- **Improved**: Updated Managed and Cad conversion engines.

- **Added**: New document formats:
- .ots - OpenDocument Spreadsheet Template
- .otp - OpenDocument Presentation Template
- .dcm - Digital Imaging and Communications in Medicine (DICOM)
- .webp - WebP Image
- .djvu - Deja Vu (DjVu) Image

- **Improved**: Added more conversions to XPS and SVG formats.