Version 2.7.5
## Version 2.7.5 - May 2, 2017

- **Fixed:** Sometimes DocumentViewer's vertical scrollbar's bottom
button was not visible.

- **Fixed:** Especially on IE, DocumentViewer's pan tool (hand cursor)
was overlapping scrollbar so it made the user
think the hotspot was on the scrollbar but when the user moved the mouse
the user was actually still panning the
document because the hotspot of the hand cursor was still on the
document and not on the scrollbar.
This confused the user thinking into clicking on scrollbar caused
"jumping back to the first page of the document".
Now, when the user approaches a scrollbar, the hand cursor will be
changed to default cursor and panning too near
the scrollbar will be disabled. This way panning and scrolling actions
will never be confused.

- **Fixed:** When printing on IE, an extra blank page was being added at
the end.

- **Improved:** Conversion and viewing accuracy for Presentation and
Imaging formats.

- **Added:** More client-side events: ClientLoad, ClientError,
ClientDocumentLoad, ClientPageComplete.
Updated example projects with the new "Events" sample demonstrating all
client-side events.