Version 3.3.0
## Version 3.3.0 - August 9, 2017

- **Added:** DocumentUltimateWebConfiguration.CacheKeepVariations option
which specifies whether to keep variations
of document conversion results. The default is false. For instance when
you change watermark options, both
Pdf and Xpz outputs are regenerated. By default the Pdf and Xpz outputs
are replaced in cache when watermarks
are changed. If you set this property to true, the Pdf and Xpz outputs
will not be replaced but all variations
will be kept in cache instead. This can be useful especially when you
want different outputs for different users.
So this way, for instance you can have different watermarked variations
for each different user.

- **Fixed:** Conversion of a large document could fail upon first view
in DocumentViewer due to the default
ASP.NET timeout (110 seconds). Increased timeout so that conversions
that take long time are completed
successfully.

- **Fixed:** Made sure DocumentViewer "please wait" dialog is hidden
when there is an error. Fixed show/hide
dialog timing.

- **Improved:** Now DocumentViewerHandler methods will use cacheInfoKey
parameter for identifying a request.
DownloadFileName, DownloadEnabled and DownloadAsPdfEnabled properties
are now persisted in DocumentViewerState
instead of DocumentCacheInfo because there properties belong to a
specific DocumentViewer instance rather than
to global scope.

- **Improved:** ControlBase.StateId will not be random and will change
less often. It is now calculated according
to the url and control ID combination. This way more handler methods can
make use of browser caching across
different sessions.