Version 3.6.0
## Version 3.6.0 - October 12, 2017

- **Fixed:** The below error which happened in some environments:

The type initializer for '_UXg' threw an exception.
---> Could not load file or assembly 'PortableEngine, Version=2.7.3.0,
Culture=neutral, PublicKeyToken=a6f3cafa178e6038' or one of its
dependencies.

The error happened either:

- When ASP.NET impersonation was used via <identity impersonate="true"
/> tag in web.config

- When Glimpse library (diagnostics & insights library) was used in the
project.

In both cases, the error is now fixed.

- **Fixed:** Viewing some TIFF files caused crashing of IIS worker
process.

- **Fixed:** JS error when embedding DocumentViewer via iframe into a
page of different origin in Safari browser.

- **Added:** DocumentViewer.HighlightedColor property for overriding the
default color (yellow) of the highlights
when you pass in a set of keywords.