Version 3.11.5
## Version 3.11.5 - March 6, 2018

- **Fixed:** In cache folder generated PDF files were left locked (the
file was not closed after converting to XPZ).
This would prevent the deletion of the folder and maybe would cause some
other strange side-effects.

- **Added:** WordProcessingInputOptions.UpdateFields property which
indicates if fields should be updated before
the DOC/DOCX document is rendered/converted. The default is true (also
the behaviour in previous versions).
When set to true, it will mimic default "Save as PDF" or "Print"
behaviour in Microsoft Word.
For example, if you see "Error! Reference Source Not Found" messages
instead of clickable links
in the rendered document, you can set this to false (until you actually
fix reference errors in your document).