Version 3.13.0
## Version 3.13.0 - March 16, 2018

- **Fixed:** Some issues with the search features that were added in
3.12.0:

- "Any words" search did not find results when keywords had hyphens
in-between.

- "Any words" search would hang (infinite loop) when one of the words
was a single hyphen.

- A subsequent search in the viewer (without reloading the document)
with new criteria where
no results are returned did not clear any of the highlights (yellow and
orange both) from
the previous successful search.

- **Improved:** Updated DocumentViewer icons with better ones and
hopefully fixed "gibberish font icon" issue
which was supposed to be fixed back in 3.11.0 but was still being
observed though not frequently as before.

- **Fixed:** When using DocumentSource with a byte array or stream, if
the input document was PDF,
Download action would fail with error "downloadError". This was because
the original copy in DocumentCache was
being replaced with the generated one.

- **Added:** DocumentViewer.DocumentHandlerParameters property for a
better solution to pass multiple parameters
to your IDocumentHandler implementation. Added DocumentHandlerParameters
parameter to the methods of
IDocumentHandler interface so you should update GetInfo and OpenRead
signatures in your implementation.
Refer to DocumentViewer.DocumentHandlerParameters property in the
updated docs for more details.