Version 3.15.5
## Version 3.15.5 - April 7, 2018

- **Fixed:** In some cases, ASP.NET's UrlRoutingModule was still taking
over the handlers (HTTP 404) after
the new own routing was added in v3.15.0. This is now completely fixed.
For reference, it was avoidable with v3.15.0 via adding the below line
at the start of RegisterRoutes method
in RouteConfig.cs (of MVC project):

routes.IgnoreRoute("{resource}.ashx/{*pathInfo}")

Now, this is not required anymore starting with this version.

- **Added:** New property DocumentViewer.DocumentFormat. By default, the
format is determined from the file extension
but you can use this property when you don't have an extension or when
you need to override the format determined
from the extension. Refer to DocumentViewer.DocumentFormat property in
the updated docs for more details.

- **Added:** New property DocumentViewer.DocumentLocation. By default,
the file is considered on disk but you can use
this property to load files from any of the supported file systems like
UNC paths with password, Amazon S3
and Azure Blob. Refer to DocumentViewer.DocumentLocation property in the
updated docs for more details.

- **Improved:** Newer versions of Cad formats (AutoCad 2018) is now
supported.