Version 3.16.0
## Version 3.16.0 - April 20, 2018

- Improved: Improved stability, accuracy and restored conversion
performance (was slower since v3.14.0) for
WordProcessing and Presentation formats.

- Improved: Improved stability, accuracy for Spreadsheet and
ProjectManagement formats.

- **Added:** New HtmlOutputOptions class for specifying if the images
and/or css should be exported as separate files.
By default these resources were embedded to produce a single html file,
now it's possible to export separate files.
Currently it works when converting to Html from WordProcessing,
Spreadsheet and Presentation formats.

- **Fixed:** DocumentViewer or DocumentConverter could not load most of
the .svg files. It failed with the
"The image is not a raster image" error.

- **Changed:** Removed DocumentFormat.Xml (SpreadsheetML - Excel 2003
XML Spreadsheet) as .xml extension is
widely used for text files. So now by default .xml files will be treated
as plain text files and DocumentViewer
will be able to display them.

- **Fixed:** When passing the same OutputOptions instance to
DocumentConverter instance multiple times, it was using
the initial detected format (from file extension) for consecutive
conversions. So encapsulated
input and output changing values especially Format property in new
InputContext and OutputContext classes.